chore(deps): update dependencies to their latest releases
`rand` 0.9 -> 0.10 moves `random_range` and `choose` from `Rng` to the new
`RngExt` trait, so the call sites import that instead. `atoi` 2.0 -> 3.1 needs
no change. Both are internal; the public API is unchanged.

`serial_test` stays on 3.x: its 4.0 requires Rust 1.93, above the crate's 1.88
MSRV.

The `*_long_polling_pubsub` examples passed `lpop`'s count as a `usize` where
the signature takes a `u32`, and no CI feature set builds `web-examples`, so
they had stopped compiling.

+++ examples/actix_long_polling_pubsub.rs
@@ -74,7 +74,7 @@ async fn get_messages_from_queue(
     redis: &Client,
     channel: &str,
 ) -> Result<Vec<String>, ServiceError> {
-    Ok(redis.lpop(channel, i32::MAX as usize).await?)
+    Ok(redis.lpop(channel, u32::MAX).await?)
 }
 
 #[post("/{channel}")]
